refactor: consolidate client management to Service struct and assign functions to struct. Consolidate mevcommit packages into service folder too
